requirements.txt
autopep8==1.5.4
bcrypt==3.2.0
certifi==2019.11.28
cffi==1.14.3
crc8==0.1.0
cryptography==3.1.1
docopt==0.6.2
future==0.18.2
httplib2==0.17.0
paramiko==2.7.2
pycodestyle==2.6.0
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.15.0
python-telegram-bot==12.3.0
pytunneling==0.0.2
six==1.15.0
sshtunnel==0.1.5
toml==0.10.2
tornado==6.0.3
wincertstore==0.2
